+++ client/lib/error-logger/setup-error-logger.js
